requirements.txt
anyio==3.3.4
asgiref==3.4.1
astroid==2.9.3
bandit==1.7.2
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
cryptography==36.0.1
dnspython==2.1.0
dodgy==0.2.1
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.70.0
flake8==4.0.1
flake8-polyfill==1.0.2
gitdb==4.0.9
GitPython==3.1.26
greenlet==1.1.2
h11==0.12.0
httptools==0.2.0
idna==3.3
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.2
lazy-object-proxy==1.7.1
MarkupSafe==2.0.1
mccabe==0.6.1
mypy==0.931
mypy-extensions==0.4.3
orjson==3.6.4
passlib==1.7.4
pbr==5.8.1
pep8-naming==0.10.0
platformdirs==2.5.0
prospector==1.6.0
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.8.2
pydocstyle==6.1.1
pyflakes==2.3.1
pylint==2.12.2
pylint-celery==0.3
pylint-django==2.5.0
pylint-flask==0.6
pylint-plugin-utils==0.7
python-dotenv==0.19.1
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==5.4.1
requests==2.26.0
requirements-detector==0.7
rsa==4.8
setoptconf-tmp==0.3.1
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
snowballstemmer==2.2.0
SQLAlchemy==1.4.31
starlette==0.16.0
stevedore==3.5.0
toml==0.10.2
tomli==2.0.1
typing-extensions==3.10.0.2
ujson==4.2.0
urllib3==1.26.7
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==10.0
wrapt==1.13.3
